Update README.md
is a basic prototype which could
be used as a receptionist at the hospitals which takes
data and adds it in a hospital database. This project is highly user
friendly and could be utilized by disabled also after some modifications like
using Braille keyboard for the visually impaired person. Also, this project
speaks to them(using playsound module) about what is to be entered and also warns them if their
symptoms match with COVID 19. Also, this project is bilingual and right now it works in Hindi
and English.
To use this project- 
1. You must have Python3 and MySQL installed in your system.
2. Install MySQL-Python connector using - pip install mysql-connector-python in command prompt.
3. Download all the files in a single folder and then update the MySQL password in the 'S.A.R.T.H.I.py' file according to your system.